Bug #76468 review follow-up (Blocking 1): eliminate the chunk-split that excluded real app code from elements.js
ngovus-inetsoft found the JS-glob fix alone wasn't sufficient: the
esbuild application builder code-splits EmbedChartComponent's ~1.9MB of
code into a separate chunk file, because embed-chart.routes.ts's shared
embedChartRoutes defines it via a lazy loadComponent() -- even though
main-elements.ts already imports EmbedChartComponent statically and uses
it immediately (createCustomElement). The gulp-concatenated elements.js
never referenced or copied that chunk, so it opened with a dangling
top-level `import` and the actual app code was missing regardless of the
JS-glob fix.

esbuild treats a dynamic import() as a hard chunk-split point for any
build that pulls in the file containing it, even from an unused export --
so simply adding an eager route alongside the lazy one in the same file
wasn't enough (confirmed: still produced the same 1.9MB chunk). Split the
matcher, providers, and the two route variants (lazy vs. eager) into
separate files, so main-elements.ts's module graph never includes the
file with the `import()` at all. Portal's own lazy-loading of
embed-chart.routes (via app.routes.ts's loadChildren) is untouched --
verified a dev build of portal still produces separate embed-chart.component
and embed-chart.routes lazy chunks, unchanged.

Verified: elements/viewer-element builds now produce zero chunk-*.js
files (previously one ~1.9MB chunk for elements); the gulp-concatenated
elements.js has zero dangling `import` statements and parses successfully
as a classic script (`new Function(src)`, which correctly forces
non-module parse semantics -- `node --check` gives a false pass here
since this repo's package.json has "type": "module").

Does NOT fix Blocking 2 (the import.meta.url Worker reference, also a
classic-script SyntaxError) -- filed as Bug #76473, a distinct problem
needing a real bundler pass (esbuild --format=iife) rather than a source
change, per discussion with the user/lead.
