+++ web/projects/portal/src/app/embed/chart/embed-chart.routes.ts
@@ -15,157 +15,22 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import { Injector, Optional } from "@angular/core";
-import { Routes, UrlMatchResult, UrlSegment } from "@angular/router";
-import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
-import { DndService } from "../../common/dnd/dnd.service";
-import { VSDndService } from "../../common/dnd/vs-dnd.service";
-import { FullScreenService } from "../../common/services/full-screen.service";
-import { UIContextService } from "../../common/services/ui-context.service";
-import { ViewsheetClientService } from "../../common/viewsheet-client";
-import { ChartService } from "../../graph/services/chart.service";
-import {
-   ComposerToken,
-   ContextProvider,
-   EmbedAssemblyContextProviderFactory
-} from "../../vsobjects/context-provider.service";
-import { RichTextService } from "../../vsobjects/dialog/rich-text-dialog/rich-text.service";
-import { VSChartService } from "../../vsobjects/objects/chart/services/vs-chart.service";
-import { DataTipService } from "../../vsobjects/objects/data-tip/data-tip.service";
-import { PopComponentService } from "../../vsobjects/objects/data-tip/pop-component.service";
-import { MiniToolbarService } from "../../vsobjects/objects/mini-toolbar/mini-toolbar.service";
-import { ShowHyperlinkService } from "../../vsobjects/show-hyperlink.service";
-import { CheckFormDataService } from "../../vsobjects/util/check-form-data.service";
-import { VSTabService } from "../../vsobjects/util/vs-tab.service";
-import { ModelService } from "../../widget/services/model.service";
-import { ScaleService } from "../../widget/services/scale/scale-service";
-import { VSScaleService } from "../../widget/services/scale/vs-scale.service";
-import {
-   DialogService,
-   ViewerDialogServiceFactory
-} from "../../widget/slide-out/dialog-service.service";
-import { SlideOutService } from "../../widget/slide-out/slide-out.service";
-import { AdhocFilterService } from "../../vsobjects/objects/data-tip/adhoc-filter.service";
+import { Routes } from "@angular/router";
 import { canDeactivateGuard } from "../../common/services/can-deactivate-guard.service";
+import { EMBED_CHART_URL_MATCHER } from "./embed-chart-url-matcher";
+import { EMBED_CHART_ROUTE_PROVIDERS } from "./embed-chart.route-providers";
 
-export function EMBED_CHART_URL_MATCHER(url: UrlSegment[]): UrlMatchResult {
-   let result: UrlMatchResult = null;
-
-   if(url && url.length > 0) {
-      const params: { [name: string]: UrlSegment } = {};
-      result = {
-         consumed: url,
-         posParams: params
-      };
-
-      if(url.length > 0) {
-         let assetScope: UrlSegment = null;
-         let assetOwner: UrlSegment = null;
-         let assetPath: string = null;
-         let assetId: string = null;
-
-         if(url[0].path === "global") {
-            assetScope = url[0];
-
-            if(url.length > 1) {
-               assetPath = url.slice(1, url.length - 1).map((s) => s.path).join("/");
-               assetId = `1^128^__NULL__^${assetPath}`;
-            }
-         }
-         else if(url[0].path === "user") {
-            assetScope = url[0];
-
-            if(url.length > 1) {
-               assetOwner = url[1];
-
-               if(url.length > 2) {
-                  assetPath = url.slice(2, url.length - 1).map((s) => s.path).join("/");
-                  assetId = `4^128^${assetOwner.path}^${assetPath}`;
-               }
-            }
-         }
-         else {
-            assetId = url.slice(0, url.length - 1).map((s) => s.path).join("/");
-            const match = /^([14])+\^128\^([^^]+)\^(.+)$/.exec(assetId);
-
-            if(match) {
-               if(match[1] == "1") {
-                  assetScope = new UrlSegment("global", {});
-               }
-               else {
-                  assetScope = new UrlSegment("user", {});
-                  assetOwner = new UrlSegment(match[2], {});
-               }
-
-               assetPath = match[3];
-            }
-         }
-
-         if(assetScope) {
-            params.assetScope = assetScope;
-         }
-
-         if(assetOwner) {
-            params.assetOwner = assetOwner;
-         }
-
-         if(assetPath) {
-            params.assetPath = new UrlSegment(assetPath, {});
-         }
-
-         if(assetId) {
-            params.assetId = new UrlSegment(assetId, {});
-         }
-
-         params.assemblyName = new UrlSegment(url[url.length - 1].path, {});
-      }
-   }
-
-   return result;
-}
-
+// Lazily loaded route used by the full portal app (app.routes.ts's `loadChildren` on this whole
+// module). NOTE: nothing in the "elements" web-component bundle (main-elements.ts) may import
+// this file -- see embed-chart.routes-eager.ts for the equivalent eager route that bundle uses
+// instead, and why (Bug #76468: this file's `import()` below is a hard chunk-split point for
+// esbuild in *any* build that pulls this module in, whether or not embedChartRoutes itself ends
+// up used by that build).
 export const embedChartRoutes: Routes = [
    {
       loadComponent: () => import("./embed-chart.component").then(m => m.EmbedChartComponent),
       canDeactivate: [canDeactivateGuard],
       matcher: EMBED_CHART_URL_MATCHER,
-      providers: [
-         DataTipService,
-         PopComponentService,
-         MiniToolbarService,
-         VSChartService,
-         SlideOutService,
-         UIContextService,
-         CheckFormDataService,
-         ShowHyperlinkService,
-         VSTabService,
-         RichTextService,
-         FullScreenService,
-         AdhocFilterService,
-         {
-            provide: DialogService,
-            useFactory: ViewerDialogServiceFactory,
-            deps: [NgbModal, SlideOutService, Injector, UIContextService]
-         },
-         {
-            provide: DndService,
-            useClass: VSDndService,
-            deps: [ModelService, NgbModal, ViewsheetClientService]
-         },
-         {
-            provide: ScaleService,
-            useClass: VSScaleService
-         },
-         {
-            provide: ContextProvider,
-            useFactory: EmbedAssemblyContextProviderFactory,
-            deps: [[new Optional(), ComposerToken]]
-         },
-         {
-            provide: ChartService,
-            useExisting: VSChartService
-         },
-         NgbModal
-      ]
+      providers: EMBED_CHART_ROUTE_PROVIDERS
    }
 ];
